Support config to override default monsterDir so that we can have aggregated configs

+++ src/sqlJobStore.js
@@ -129,6 +129,6 @@ export class SqlJobStore {
     }
 
     async shutdown() {
-        await this.db.destroy();
+        this.db && await this.db.destroy();
     }
 }
