+++ src/sqlStateStore.js
@@ -76,6 +76,6 @@ export class SqlStateStore {
     }
 
     async shutdown() {
-        await this.db.destroy();
+        this.db && await this.db.destroy();
     }
 }
